Pull the large `Database.ts` file into some smaller chunks
This change introduces the concept of "internal" files. Files ending
with `.internal.{ts,tsx}` will now be ignored from the barrel file, so
will not be easily accessed by consumers of the library, while still
allowing sharing between modules internally.
