+++ barrelsby.json
@@ -1,5 +1,9 @@
 {
   "directory": "src",
   "delete": true,
-  "exclude": ["^.*/__.+__/.+\\.tsx?$", "^.+\\.(spec|test)\\.tsx?$"]
+  "exclude": [
+    "^.*/__.+__/.+\\.tsx?$",
+    "^.+\\.(spec|test)\\.tsx?$",
+    "^.+\\.internal\\.tsx?$"
+  ]
 }
